Moved read code from `Connection` to `RecvQueue`
